pyproject.toml
[project]
name = "sales-automation-agent"
version = "0.1.0"
description = "AI-powered sales automation agent using Google ADK with MCP integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.10.0",
    "google-auth>=2.40.3",
    "google-cloud-aiplatform[adk,agent-engines]>=1.88.0",
    "mcp>=1.12.4",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "aiofiles>=23.2.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dateutil>=2.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
